// oneHub/datart-versions/datart-1.0.0-rc.1/frontend/src/app/pages/MainPage/pages/ViewPage/slice/types.ts
export type DatabaseSchema = {
  dbName: string;
  tables: Array<{
    primaryKeys: string[];
    tableName: string;
    columns: Array<{
      fmt: string;
      foreignKeys: Array<{ column: string; database: string; table: string }>;
      name: string;
      type: string;
    }>;
  }>;
};
// ...
export interface StructViewQueryProps {
  table: Array<string>;
  columns: Array<string>;
  joins: Array<JoinTableProps>;
}

export interface JoinTableProps {
  table?: Array<string>;
  joinType?: StructViewJoinType;
  columns?: Array<string>;
  conditions?: Array<{ left: Array<string>; right: Array<string> }>;
}